Make atoms arg const and make constrain() not alter the atoms
The contrain() method altered the atom's value in-place as a performance optimization. However, the non-const argument is not ideal since it can result in problems somewhere else. The only solution for those problems is to copy the atoms before calling these methods, which basically makes the optimization later on useless.

So while there is a potential performance benefit to clamping in-place, I don't think it outweights the downsides.