Move timer implementation to timer_base
This is necessary for the next commit: we want to have a pointer from a different struct back to our timer class. However, if the timer class is templated, we cannot do that properly. Therefore, we move most of the code of the timer class to timer_base, which is a non-template class.