Pass a Max object to clock_new in order for the clock to use the proper scheduler
clock_new’s first argument expects a pointer to a Max object. In this context, ‘this’ is a pointer to our C++ timer/timer_base class, which is not a proper Max object.

To make clock_new behave correctly, there’s now a timer_impl Max object that is passed to clock_new. The timer_impl then calls into timer_base.